Adding symlink to index.html
When a user navigates to https://emmasax1.github.io, it currently just shows the README.md, which isn't very helpful for users because that's not the actual site. This is because I want the first page a user sees to be my `home.html`, not the `index.html` (at least that's what I want the URL to show).

But GitHub still needs the `index.html` file to exist, and for that to be the first thing a user sees. So, make an `index.html` file, but then just symlink it to `home.html`, so essentially both exist and are exactly the same.